Add compatibility header, refactor IrrigoSystem and MenuSystem libraries
- Introduced Compatibility.h for min/max macros during unit testing.
- Created library.json files for IrrigoSystem and MenuSystem with dependencies.
- Refactored IrrigoSystem to use composition over inheritance, initializing components in the constructor.
- Updated MainSystem to accept dependencies via constructor and removed dynamic memory allocation.
- Modified SystemManager and its interface to include update functionality.
- Implemented IMenuSystem interface in MenuSystem and updated its methods accordingly.
- Added mock classes for testing, including MockBlinker, MockDisplay, MockInputDevice, MockMainSystem, MockMenuObjects, MockRtc, and MockSystemManager.
- Created unit tests for basic functionality in test files.
- Cleaned up platformio.ini for better organization and added include paths.

+++ lib/IrrigoSystem/src/core/SystemManager.cpp
@@ -1,4 +1,5 @@
 #include "SystemManager.h"
+#include <Compatibility.h>
 
 SystemManager::SystemManager(IMainSystem *mainSystem)
     : mainSystem(mainSystem)
@@ -100,7 +101,7 @@ void SystemManager::incrementYear()
 {
     IRtc *rtc = mainSystem->getRTC();
     RtcDateTime now = rtc->now();
-    RtcDateTime newDateTime = {now.year + 1, now.month, now.day, now.hour, now.minute, now.second};
+    RtcDateTime newDateTime = {(uint16_t)(now.year + 1), now.month, now.day, now.hour, now.minute, now.second};
     rtc->adjust(newDateTime);
 }
 
@@ -108,7 +109,7 @@ void SystemManager::decreaseYear()
 {
     IRtc *rtc = mainSystem->getRTC();
     RtcDateTime now = rtc->now();
-    RtcDateTime newDateTime = {now.year - 1, now.month, now.day, now.hour, now.minute, now.second};
+    RtcDateTime newDateTime = {(uint16_t)(now.year - 1), now.month, now.day, now.hour, now.minute, now.second};
     rtc->adjust(newDateTime);
 }
